fix: fix int enum
