+++ client/src/ee/types/types.ts
@@ -158,13 +158,16 @@ export interface LicenseInfo {
 
 /**
  * MessageParam
- * Message param model agnostic to providers.
+ * Message parameter model.
  */
 export interface MessageParam {
-  /** Role */
-  role: string;
-  /** Content */
-  content: (AudioPart | TextPart | ImagePart | ToolCall)[];
+  /** Message */
+  message: string;
+  /**
+   * Timestamp
+   * @format date-time
+   */
+  timestamp: string;
 }
 
 /**
@@ -220,10 +223,10 @@ export interface SpanMoreDetails {
  * License tier enum.
  */
 export enum Tier {
-  VALUE_0 = 0,
-  VALUE_1 = 1,
-  VALUE_2 = 2,
-  VALUE_3 = 3,
+  FREE = 0,
+  PRO = 1,
+  TEAM = 2,
+  ENTERPRISE = 3,
 }
 
 /** ValidationError */
@@ -234,56 +237,4 @@ export interface ValidationError {
   msg: string;
   /** Error Type */
   type: string;
-}
-
-/**
- * _AudioPart
- * Image part model.
- */
-export interface AudioPart {
-  /** Type */
-  type: "audio";
-  /** Media Type */
-  media_type: string;
-  /** Audio */
-  audio: string;
-}
-
-/**
- * _ImagePart
- * Image part model.
- */
-export interface ImagePart {
-  /** Type */
-  type: "image";
-  /** Media Type */
-  media_type: string;
-  /** Image */
-  image: string;
-  /** Detail */
-  detail: string | null;
-}
-
-/**
- * _TextPart
- * Text part model.
- */
-export interface TextPart {
-  /** Type */
-  type: "text";
-  /** Text */
-  text: string;
-}
-
-/**
- * _ToolCall
- * Image part model.
- */
-export interface ToolCall {
-  /** Type */
-  type: "tool_call";
-  /** Name */
-  name: string;
-  /** Arguments */
-  arguments: object;
 }
