+++ lilypad/ee/server/require_license.py
@@ -9,9 +9,10 @@ from uuid import UUID
 
 from fastapi import Depends, HTTPException, Request, status
 
-from ee import LicenseError, LicenseInfo, LicenseValidator, Tier
+from ee import LicenseInfo, LicenseValidator, Tier
 
 from ...ee.server import HOST_NAME
+from ...exceptions import LicenseError
 from ...server._utils import get_current_user
 from ...server.exceptions import LilypadForbiddenError
 from ...server.schemas import UserPublic
